+++ Plugins/Crafting/src/main/java/com/tagnumelite/projecteintegration/plugins/crafting/PluginArtisanWorktables.java
@@ -23,9 +23,9 @@ public class PluginArtisanWorktables extends APEIPlugin {
 
     @Override
     public void setup() {
-        for (EnumTier tier : EnumTier.values()) {
-            if (!ModuleWorktablesConfig.isTierEnabled(tier)) {
-                for (String name : ArtisanAPI.getWorktableNames()) {
+        for (String name : ArtisanAPI.getWorktableNames()) {
+            for (EnumTier tier : EnumTier.values()) {
+                if (!ModuleWorktablesConfig.isTierEnabled(tier)) {
                     addMapper(new WorktableMapper(name, tier));
                 }
             }
@@ -37,7 +37,7 @@ public class PluginArtisanWorktables extends APEIPlugin {
         private final EnumTier tier;
 
         public WorktableMapper(String name, EnumTier tier) {
-            super(name + '_' + tier.getName(), "Enable mapper for worktable " + name + ' ' + tier.getName() + '?');
+            super(name + ' ' + tier.getName(), "Enable mapper for worktable " + name + ' ' + tier.getName() + '?');
             this.name = name;
             this.tier = tier;
         }
@@ -46,7 +46,6 @@ public class PluginArtisanWorktables extends APEIPlugin {
         public void setup() {
             List<IArtisanRecipe> recipe_list = new ArrayList<>();
             recipe_list = ArtisanAPI.getWorktableRecipeRegistry(name).getRecipeListByTier(tier, recipe_list);
-
 
             for (IArtisanRecipe recipe : recipe_list) {
                 List<SizedIngredient> ingredients = new ArrayList<>();
@@ -61,22 +60,18 @@ public class PluginArtisanWorktables extends APEIPlugin {
                 ArrayList<Object> outputs = new ArrayList<>();
                 for (OutputWeightPair output : recipe.getOutputWeightPairList()) {
                     outputs.add(output.getOutput().toItemStack());
-                    //addRecipe(output.getOutput().toItemStack(), fluid_input, ingredients.toArray());
                 }
 
                 if (recipe.getSecondaryOutputChance() >= 1f) {
                     outputs.add(recipe.getSecondaryOutput().toItemStack());
-                    //addRecipe(recipe.getSecondaryOutput().toItemStack(), fluid_input, ingredients.toArray());
                 }
 
                 if (recipe.getTertiaryOutputChance() >= 1f) {
                     outputs.add(recipe.getTertiaryOutput().toItemStack());
-                    //addRecipe(recipe.getTertiaryOutput().toItemStack(), fluid_input, ingredients.toArray());
                 }
 
                 if (recipe.getQuaternaryOutputChance() >= 1f) {
                     outputs.add(recipe.getQuaternaryOutput().toItemStack());
-                    //addRecipe(recipe.getQuaternaryOutput().toItemStack(), fluid_input, ingredients.toArray());
                 }
 
                 addRecipe(outputs, fluid_input, ingredients.toArray());